Adding Props, Materials, And Textures
I create a new folder called props, and adding the props as listed above, I also import the texture files in the "Textures" folder. To read the texture files, the props meshes needs a material. So, i also create a folder called props material and create a material for each one of the meshes.